+++ owner/src/main/java/com/mdh/owner/global/security/LoginService.java
@@ -2,8 +2,8 @@ package com.mdh.owner.global.security;
 
 import com.mdh.common.user.persistence.UserRepository;
 import com.mdh.owner.global.security.controller.SignUpRequest;
+import com.mdh.owner.global.security.session.CustomUser;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
