+++ owner/src/main/java/com/mdh/owner/global/security/session/CurrentUser.java
@@ -1,4 +1,4 @@
-package com.mdh.owner.global.security;
+package com.mdh.owner.global.security.session;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
