+++ owner/src/main/java/com/mdh/owner/global/security/session/CurrentUserHandlerMethodArgumentResolver.java
@@ -1,8 +1,7 @@
-package com.mdh.owner.global.security;
+package com.mdh.owner.global.security.session;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
