+++ owner/src/main/java/com/mdh/owner/reservation/read/presentation/OwnerReservationReadController.java
@@ -3,8 +3,8 @@ package com.mdh.owner.reservation.read.presentation;
 import com.mdh.owner.global.ApiResponse;
 import com.mdh.common.reservation.domain.ReservationStatus;
 import com.mdh.common.reservation.persistence.dto.OwnerShopReservationInfoResponse;
-import com.mdh.owner.global.security.CurrentUser;
-import com.mdh.owner.global.security.UserInfo;
+import com.mdh.owner.global.security.session.CurrentUser;
+import com.mdh.owner.global.security.session.UserInfo;
 import com.mdh.owner.reservation.read.application.OwnerReservationReadService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
